% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ai_mean_temperature}
\alias{ai_mean_temperature}
\title{Aleutian Islands summer bottom trawl survey mean temperature (ai_mean_temperature)}
\format{
data.frame
\describe{
     \item{YEAR}{Year}
     \item{SUBAREA}{Aleutian Islands Ecosystem Status Report subarea (Western/Central/Eastern).}
     \item{MEAN_GEAR_TEMPERATURE}{Mean gear temperature (i.e. bottom temperature)}
     \item{MEAN_SURFACE_TEMPERATURE}{Mean sea surface temperature.}
     \item{SE_GEAR_TEMPERATURE}{Standard error of mean gear temperature.}
     \item{SE_SURFACE_TEMPERATURE}{Standard error of mean surface temperature.}
     \item{LAST_UPDATE}{Date when data set was last updated.}
}
}
\source{
\url{https://www.fisheries.noaa.gov/contact/groundfish-assessment-program}
}
\usage{
ai_mean_temperature
}
\description{
Data frame containing mean temperatures from AFSC/RACE/GAP summer bottom trawl surveys of the Aleutian Islands for western, central, and eastern Aleutian Islands Ecosystem Status Report subareas.
}
\keyword{datasets}
