% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fixed_depth_var_bt.R
\name{calc_fixed_depth_var_bt}
\alias{calc_fixed_depth_var_bt}
\title{Estimate value of a profile variable at a fixed depth}
\usage{
calc_fixed_depth_var_bt(depth, var, ref_depth)
}
\arguments{
\item{depth}{Numeric vector of depth values}

\item{var}{vector of variables}

\item{ref_depth}{Reference depth for estimating 1L numeric vector}
}
\description{
Algorithm used since 2013 to estimate temperature at 1-m (RACE_DATA.TEMPERATURE_METHOD.TEMPERATURE_METHOD_ID = 14) from temperature-depth recorders deployed on RACE bottom trawl surveys. Returns the variable value corresponding with the first depth value at or shallower than the reference depth. If no depth values are equal to the reference depth, the variable value is estimated by interpolating from the variable values above and below the reference depth.
}
\author{
Sean Rohan
}
