% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_spatraster.R
\name{cluster_spatraster}
\alias{cluster_spatraster}
\title{Hierarchical clustering on SpatRaster objects}
\usage{
cluster_spatraster(
  x,
  scale_vars = TRUE,
  dist_method = "euclidean",
  hclust_method = "complete"
)
}
\arguments{
\item{x}{A numeric data frame or matrix where rows represent individual
observations (e.g., cities, subjects) and columns represent variables measured
over time (e.g., months, days).}

\item{scale_vars}{A logical value. If `TRUE` (the default), the data for each
row is standardized (scaled) before clustering.}

\item{dist_method}{The distance measure to be used by `dist()`. Defaults to
`"euclidean"`. See help documentation for `dist()` for additional options.}

\item{hclust_method}{The agglomeration method to be used by `hclust()`.
Defaults to `"complete"` for complete linkage clustering. See help documentation for `hclust()` for additional options.}
}
\value{
A list containing three objects:
  \item{mat}{The processed (and possibly scaled and transposed) matrix used for the analysis.}
  \item{dist}{The distance object (`dist`) calculated from the processed matrix.}
  \item{clust}{The hierarchical clustering object (`hclust`) produced by the analysis.}
}
\description{
This function performs hierarchical clustering on a terra 'SpatRaster' object. It is intended for use on rasters where each layer represents an observation in time (e.g. annual temperature observations).
}
\details{
The analysis pipeline consists of several steps:
1. The input `x` is converted to a matrix and any rows with `NA` values are removed.
2. If `scale_vars = TRUE`, each row is standardized to have a mean of 0 and a
   standard deviation of 1. This focuses the analysis on the pattern of
   the values over time, rather than their magnitude.
3. A distance matrix is calculated between the time points using the method
   specified by `dist_method`.
4. Hierarchical clustering is performed using the method specified by `hclust_method`.
5. A dendrogram of the clustering result is plotted.
}
\examples{
# Hierarchical clustering on EBS shelf bottom temperature SpatRaster

ebs_bt <- terra::unwrap(coldpool::ebs_bottom_temperature)

cluster_spatraster(
  x = ebs_bt,
  scale_vars = TRUE,
  dist_method = "euclidean",
  hclust_method = "complete"
)
}
