% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_cpa_station_filter.R
\name{compare_cpa_station_filter}
\alias{compare_cpa_station_filter}
\title{Effect of interpolating with and without RKC resample}
\usage{
compare_cpa_station_filter(
  sel_year = 2000,
  sel_old_raster = "./data/idw_files/bt00_idw/",
  sel_idw4_geotiff = "./output/raster/idw_nmax4_2000_gear_temperature.tif",
  set_crs = "EPSG:3338",
  temp_data_path = here::here("data",
    list.files(here::here("data"))[length(list.files(here::here("data")))])
)
}
\arguments{
\item{sel_year}{Year for plot}

\item{sel_old_raster}{File path to an old raster from a year with RKC resampling}

\item{sel_idw4_geotiff}{File path to a geotiff IDW raster with nmax = 4.}

\item{set_crs}{Coordinate reference system}

\item{temp_data_path}{Filed path to the temperature data .csv}
}
\description{
Compare old (pre-2021) cold pool area calculations that included red king crab resample to new cold pool area calculations.
}
