% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_ddm_to_dd}
\alias{convert_ddm_to_dd}
\title{Convert decimal degree minutes to decimal degrees}
\usage{
convert_ddm_to_dd(x)
}
\arguments{
\item{x}{Numerical vector of decimal degree minutes.}
}
\value{
Numerical vector of decimal degrees
}
\description{
Converts latitude or longitude in decimal degree minutes, DDDMM.MMM (i.e. -17330.000) to decimal degrees, DDD.DDDDD (i.e. -173.50000).
}
