% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2raster.R
\name{data2raster}
\alias{data2raster}
\title{Interpolate data points to a masked raster}
\usage{
data2raster(
  dat,
  dat.year,
  var.col,
  lat.col,
  lon.col,
  in.crs = "+proj=longlat +datum=NAD83",
  interpolation.crs,
  cell.resolution,
  nm = Inf,
  select.region = "sebs",
  fittypes = c("nn", "idwnmax4", "idw", "exp", "sph", "bes", "gau", "cir", "mat", "ste",
    "tps"),
  bbox = NULL
)
}
\arguments{
\item{dat}{Input data frame. Must include columns for latitude, longitude, 
and variable to be interpolated.}

\item{dat.year}{Year}

\item{var.col}{Character vector denoting the name of the variable column.}

\item{lat.col}{Character vector indicating the name of the latitude column.}

\item{lon.col}{Character vector indicating the name of the longitude column.}

\item{in.crs}{Character vector (PROJ4 or WKT2/PROJ6 string) indicating the 
coordinate reference system for the data.}

\item{interpolation.crs}{Character vector (PROJ4 or WKT2/PROJ6 string) 
indicating the coordinate reference system to use for the interpolation. 
Historically EPSG:3338}

\item{cell.resolution}{Dimensions of interpolation grid cell in meters.}

\item{nm}{Maximum number of cells to use for interpolation.}

\item{select.region}{Region for interpolation as a character string. Can be 
either a string accepted by akgfmaps::get_base_layers or a simple feature 
polygon}

\item{fittypes}{vector of strings specifying which fit algorithms to use:
"nn" = nearest neighbor
"idwnmax4" = inverse distance weighting with nmax=4
"idw" = inverse distance weighting
"exp" = ordinary kriging w/ exponential VGM, 
"sph" = ordinary kriging w/ spherical VGM
"bes" = ordinary kriging w/ bessel VGM
"gau" = ordinary kriging w/ gaussian VGM
"cir" = ordinary kriging w/ circular VGM
"mat" = ordinary kriging w/ matern VGM
"ste" = ordinary kriging w/ Stein's matern VGM
"tps" = thin plate spline}

\item{bbox}{Bounding box for the interpolated grid.  Can be either a string 
("sebs", "bs.south", "nebs", "ebs", "bs.north", "bs.all") corresponding to 
specific bounds, or a 4-element named vector with values xmin, xmax, ymin, 
and ymax.}
}
\value{
raster stack of interpolated data, with one layer per input fit type
}
\description{
This function is a variant of the interpolate_variable function that allows
for user-set masking polygons and choice of fitting method.
}
