% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ebs_proj_crs}
\alias{ebs_proj_crs}
\title{CRS for eastern Bering Sea cold pool index}
\format{
A 1L character vector
}
\usage{
ebs_proj_crs
}
\description{
Character vector for North American Datum 1983 / Albers Equal Area Alaska (EPSG:3338).
}
\keyword{datasets}
