% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ebs_surface_temperature}
\alias{ebs_surface_temperature}
\title{Rasters of summer sea surface temperature in the eastern Bering Sea at 5-km resolution}
\format{
A RasterStack class with 39 layers, one for every year from 1982-2021, except for 2020 (due to survey cancellation).
#' \describe{
     \item{ste_(year)_gear_temperature}{Raster layer of temperature for a given year}
}
}
\source{
\url{https://www.fisheries.noaa.gov/contact/groundfish-assessment-program}
}
\usage{
ebs_surface_temperature
}
\description{
Summer surface temperatures in the eastern Bering Sea continental shelf survey area, calculated from interpolation of temperature data from summer bottom trawl surveys conducted by AFSC/RACE/GAP. To load ethis data set, it is necessary to first load the coldpool package or raster and sp packages. Raster layers were interpolated using ordinary kriging with Matern covariance \doi{10.1007/978-1-4612-1494-6}{(Stein, 1999)}.
}
\keyword{datasets}
