% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_anisotropy.R
\name{estimate_anisotropy}
\alias{estimate_anisotropy}
\title{Estimate 2D variogram anisotropy parameters for kriging using 10-fold cross-validation}
\usage{
estimate_anisotropy(
  x,
  variable_name,
  latitude_name,
  longitude_name,
  input_crs,
  interpolation_crs,
  kriging_formula = variable_name ~ 1,
  folds = 10,
  variogram_model,
  starting_values = NULL,
  vgm_width = NULL,
  nm = Inf,
  seed = 19673
)
}
\arguments{
\item{x}{A data.frame containing variables to be interpolated and latitude and longitude coordinates.}

\item{variable_name}{1L character vector indicating which varible to interpolate (e.g. 'TEMPERATURE')}

\item{latitude_name}{1L character vector indicating the name of the latitude column (e.g., 'LATITUDE')}

\item{longitude_name}{1L character vector indicating the name of the longitude column (e.g., 'LONGITUDE')}

\item{input_crs}{Character vector indicating the coordinate reference system for x (default = "WGS84")}

\item{interpolation_crs}{Coordinate reference system to use for interpolation.}

\item{kriging_formula}{Formula to use for kriging. Default (kriging_formula = variable_name ~ 1) is ordinary kriging without covariates. Formula for the gstat interface (see ?gstat::gstat). Covariate example: kriging_formula = variable_name ~ BOTTOM_DEPTH}

\item{folds}{1L numeric vector indicating how many folds to use for cross-validation.}

\item{variogram_model}{Character vector indicating which variogram model to use for interpolation. Valid options are exponential (exp), circular (cir), gaussian (gau), Bessel (bes), Matern (mat), or Stein's Matern (ste).}

\item{starting_values}{Optional. Starting value of anisotropy parameters to use for optimization. See: ?gstat::vgm.}

\item{vgm_width}{Optional. Width of variogram breaks.}

\item{nm}{Maximum number of nearest neighbor observations to use for interpolation.}

\item{seed}{RNG seed (set.seed(seed)) to use for anisotropy estimation based on cross-validation.}

\item{only_return_anisotropy}{For internal use. Default = FALSE}
}
\description{
Uses the coldpool::kriging_loocv() interface.
}
