% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_z_expansion.R
\name{estimate_z_expansion}
\alias{estimate_z_expansion}
\title{Estimate vertical expansion factor for 3D kriging}
\usage{
estimate_z_expansion(
  x,
  location_formula,
  kriging_formula,
  model,
  z_start,
  cv_index,
  anisotropy_parameters = c(0, 0, 0, 1, 1),
  vgm_width,
  nm = Inf,
  maxdist = Inf,
  use_for_rmse = rep(TRUE, length(cv_index)),
  z_limits = c(10, 5e+05)
)
}
\arguments{
\item{x}{data.frame containing}

\item{location_formula}{Formula to use for location argument to gstat. See gstat documentation for description of the formula interface (see ?gstat::gstat).}

\item{kriging_formula}{Formula to use for kriging. See gstat documentation for description of the formula interface (see ?gstat::gstat).}

\item{z_start}{Starting value for vertical expansion estimation}

\item{cv_index}{Index of folds for cross validation}

\item{vgm_width}{Optional.}

\item{nm}{Maximum number of nearest neighbor observations to use for interpolation.}

\item{use_for_rmse}{Logical vector indicating whether to use an observation to calculate RMSE for cross-validation.}

\item{z_limits}{Upper and lower bounds for z_expansion, for optimization.}

\item{anisotopy_parameters}{Anisotropy parameters to use for ordinary kriging as a 5L vector. See: ?gstat::vgm. If NULL and estimate_anisotropy, anisotropy is estimated.}

\item{variogram_model}{Character vector indicating which variogram model to use for interpolation. Valid options are exponential (exp), circular (cir), gaussian (gau), Bessel (bes), Matern (mat), or Stein's Matern (ste).}
}
\description{
Uses L-BFGS algorithm implemented in optim().
}
