% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get gear temperature data from all hauls, write to 'data/[date]_all_temperature_data.csv}
\usage{
get_data(channel)
}
\arguments{
\item{channel}{ODBC connection as an RODBC class}
}
\description{
Get gear temperature data from all hauls, write to 'data/[date]_all_temperature_data.csv
}
