% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goa_mean_temperature}
\alias{goa_mean_temperature}
\title{Gulf of Alaska summer bottom trawl survey mean temperature (goa_mean_temperature)}
\format{
data.frame
\describe{
     \item{YEAR}{Year}
     \item{SUBAREA}{Gulf of Alaska Ecosystem Status Report subarea (Western/Eastern).}
     \item{MEAN_GEAR_TEMPERATURE}{Mean gear temperature (i.e. bottom temperature).}
     \item{MEAN_SURFACE_TEMPERATURE}{Mean sea surface temperature.}
     \item{MEAN_100M_TEMPERATURE}{Mean temperature at 100 m}
     \item{MEAN_200M_TEMPERATURE}{Mean temperature at 200 m}
     \item{LAST_UPDATE}{Date when data set was last updated.}
}
}
\source{
\url{https://www.fisheries.noaa.gov/contact/groundfish-assessment-program}
}
\usage{
goa_mean_temperature
}
\description{
Data frame containing mean temperatures from AFSC/RACE/GAP summer bottom trawl surveys of the Gulf of Alaska for western and eastern Gulf of Alaska Ecosystem Status Report subareas.
}
\keyword{datasets}
