% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goa_surface_temperature}
\alias{goa_surface_temperature}
\title{Rasters of summer surface temperature in the Gulf of Alaska 1-km resolution}
\format{
terra PackedSpatRast
#' \describe{
     \item{year}{Raster layer of temperature for a given year}
}
}
\source{
\url{https://www.fisheries.noaa.gov/contact/groundfish-assessment-program}
}
\usage{
goa_surface_temperature
}
\description{
Summer surface temperatures in the Gulf of Alaska survey area, calculated from interpolation of temperature data from summer bottom trawl surveys conducted by AFSC/RACE/GAP. To load this data set, it is necessary to first load the coldpool package or terra.
}
\keyword{datasets}
