% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_gear_temp.R
\name{interpolate_gear_temp}
\alias{interpolate_gear_temp}
\title{Interpolate gear temperature}
\usage{
interpolate_gear_temp(temp_data_path, proj_crs, cell_resolution)
}
\arguments{
\item{temp_data_path}{Filepath to temperature csv file generated by get_data()}

\item{proj_crs}{CRS string to use for interpolation as a character vector (CRS should have units of meters)}

\item{cell_resolution}{Interpolation grid cell dimension in meters.}
}
\value{
Data frame of estimated cold pool area from multiple candidate interpolation methods.
}
\description{
This function interpolates gear temperature to calculate cold pool area using all candidate interpolation methods and outputs rasters of temperature and estimates of Cold Pool Area. Output includes the best
}
