% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation_wrapper.R
\name{interpolation_wrapper}
\alias{interpolation_wrapper}
\title{Wrapper function around interpolate_variable()}
\usage{
interpolation_wrapper(
  temp_data_path,
  proj_crs,
  cell_resolution,
  select_years = NULL,
  interp_variable,
  select_region = "sebs"
)
}
\arguments{
\item{temp_data_path}{Filepath to temperature csv file generated by get_data()}

\item{proj_crs}{CRS string to use for interpolation as a character vector (CRS should have units of meters)}

\item{cell_resolution}{Interpolation grid cell dimension in meters.}

\item{select_years}{Optional. Select years to use for interpolation. Default NULL uses all years.}

\item{select_region}{Region for interpolation as a character string. Options = "ebs", "sebs", "nbs"}
}
\value{
Data frame of estimated cold pool area from multiple candidate interpolation methods.
}
\description{
This function interpolates gear temperature to calculate cold pool area using all candidate interpolation methods and outputs rasters of temperature and estimates of Cold Pool Area. Output includes the best
}
