% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging_loocv.R
\name{kriging_loocv}
\alias{kriging_loocv}
\title{Compare interpolation methods using leave-one-out cross-validation}
\usage{
kriging_loocv(
  x,
  variable_name,
  latitude_name,
  longitude_name,
  elevation_name = NULL,
  input_crs,
  interpolation_crs,
  anisotropy_parameters = NULL,
  estimate_anisotropy = FALSE,
  anisotropy_kfold = 10,
  nm = Inf,
  vgm_width = NULL,
  kriging_formula = variable_name ~ 1,
  interpolation_methods = c("nn", "idw", "idw_nmax", "exp", "cir", "gau", "sph", "mat",
    "bes", "ste"),
  seed = 19673,
  only_return_anisotropy = FALSE
)
}
\arguments{
\item{x}{A data.frame containing variables to be interpolated and latitude and longitude coordinates.}

\item{variable_name}{1L character vector indicating which varible to interpolate (e.g. 'TEMPERATURE')}

\item{latitude_name}{1L character vector indicating the name of the latitude column (e.g., 'LATITUDE')}

\item{longitude_name}{1L character vector indicating the name of the longitude column (e.g., 'LONGITUDE')}

\item{elevation_name}{Optional. 1L character vector indicating the name of the elevation/depth column. Only used for 3D kriging.}

\item{input_crs}{Character vector indicating the coordinate reference system for x (default = "WGS84")}

\item{interpolation_crs}{Coordinate reference system to use for interpolation}

\item{estimate_anisotropy}{Logical indicating whether anisotropy should be estimated for kriging methods based on k-fold cross validation. Default = FALSE}

\item{anisotropy_kfold}{1L numeric vector indicating how many folds to use for cross-validation if estimate_anisotropy = TRUE.}

\item{nm}{Maximum number of nearest neighbor observations to use for interpolation.}

\item{vgm_width}{Optional. Width of variogram breaks.}

\item{kriging_formula}{Formula to use for kriging. Default (kriging_formula = variable_name ~ 1) is ordinary kriging without covariates. Formula for the gstat interface (see ?gstat::gstat). Covariate example: kriging_formula = variable_name ~ BOTTOM_DEPTH}

\item{seed}{RNG seed (set.seed(seed)) to use for anisotropy estimation based on cross-validation.}

\item{only_return_anisotropy}{For internal use. Defualt = FALSE}

\item{anisotopy_parameters}{Optional. Anisotropy parameters to use for ordinary kriging as a 2L vector. See: ?gstat::vgm. If NULL and estimate_anisotropy, anisotropy is estimated.}

\item{interplation_methods}{Interpolation methods to use. Valid options are nearest-neighbor, inverse distance weighting, inverse distance weighting using the closest nm stations (idw_nmax), and kriging with and exponential (exp), circular (cir), gaussian (gau), Bessel (bes), Matern (mat), or Stein's Matern (ste) variogram model.}
}
\description{
Conducts leave-one-out cross validation of a single predictor using kriging, nearest-neighbor, and inverse-distance weighting interpolation method using gstat.
}
