% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocv_2.R
\name{loocv_2}
\alias{loocv_2}
\title{Perform leave-one-out cross-validation and spatial interpolation on EBS data}
\usage{
loocv_2(
  dat,
  var.col,
  lat.col,
  lon.col,
  in.proj = "+proj=longlat +datum=NAD83",
  interp.proj = "+init=epsg:3338 +datum=NAD83 +units=m",
  scale.vars = FALSE,
  center = TRUE,
  scale = TRUE,
  nm = Inf,
  pre = NA,
  ...
)
}
\arguments{
\item{dat}{Data frame that contains latitude, longitude, and a variable to be interpolated.}

\item{var.col}{Name of the variable column that is to be interpolated}

\item{lat.col}{Name of the latitude column}

\item{lon.col}{Name of the longitude column}

\item{in.proj}{Projection for the input data}

\item{interp.proj}{Projection for the interpolation/output}

\item{scale.vars}{Should variable be scaled?}

\item{center}{Passed to \code{scale} if scale.vars == TRUE}

\item{scale}{Passed to \code{scale}  if scale.vars == TRUE}

\item{nm}{Maximum number of neighboring stations to use for interpolation.}

\item{pre}{Prefix for name of the output file. Default (NA) uses variable name from var.col}
}
\description{
Select a spatial interpolation methods using leave-one-out cross validation.
}
