% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocv_gear_temp.R
\name{loocv_gear_temp}
\alias{loocv_gear_temp}
\title{Compare bottom temperature interpolation methods using leave-one-out cross validation}
\usage{
loocv_gear_temp(temp_data_path, proj_crs)
}
\arguments{
\item{temp_data_path}{Filepath to temperature csv file generated by get_data()}

\item{proj_crs}{CRS for the project as a character vector.}
}
\description{
Wrapper function for coldpool::loocv_2(), to run leave-one-out cross validation.
}
