% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_loocv_summary.R
\name{make_loocv_summary}
\alias{make_loocv_summary}
\title{Function to make loocv summary tables.}
\usage{
make_loocv_summary(
  sel_paths = list.files("./output/loocv", full.names = TRUE),
  interp_variable = "gear_temperature"
)
}
\arguments{
\item{sel_paths}{File paths to loocv output files}

\item{interp_variable}{Name of the variable to calculate summary statistics for (i.e., "gear_temperature")}
}
\description{
Reads in output from leave-one-out cross-validation and makes tables summarizing mean, minimum, and maximum annual prediction error, and frequency of interpolation method being selected as the best.
}
