% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_raster_file_stack.R
\name{make_raster_file}
\alias{make_raster_file}
\title{Save raster to file}
\usage{
make_raster_file(x, filename, format, overwrite, layer_name)
}
\arguments{
\item{x}{A raster object}

\item{filename}{A filename where the raster will be written}

\item{format}{Format to use to write the raster (see raster::writeRaster)}

\item{overwrite}{Logical. Should existing rasters be overwritten}

\item{layer_name}{Character or numeric vector indicating the name for the single layer raster. Must be a valid name for a raster layer.}
}
\description{
Saves a raster to file using a custom name for the value column.
}
