% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nbs_ebs_bottom_salinity}
\alias{nbs_ebs_bottom_salinity}
\title{Rasters of summer salinity (PSS-78) in the eastern and northern Bering Sea at 5-km resolution}
\format{
A PackedSpatRast class with five layers.
#' \describe{
     \item{ste_(year)_gear_temperature}{Raster layer of salinity (PSS-78) for a given year}
}
}
\source{
\url{https://www.fisheries.noaa.gov/contact/groundfish-assessment-program}
}
\usage{
nbs_ebs_bottom_salinity
}
\description{
Summer salinity (PSS-78) in the full eastern Bering Sea survey area, calculated from interpolation of data from summer bottom trawl surveys conducted by AFSC/RACE/GAP. To load this data set, it is necessary to first load the coldpool package or terra package. Raster layers were interpolated using ordinary kriging with Matern covariance \doi{10.1007/978-1-4612-1494-6}{(Stein, 1999)}.
}
\keyword{datasets}
