% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loocv_rmse.R
\name{plot_loocv_rmse}
\alias{plot_loocv_rmse}
\title{Function to make violin plots of MSE}
\usage{
plot_loocv_rmse(
  sel_paths = dir("./output/loocv", full.names = TRUE),
  y_lab = NULL,
  make_plot = TRUE,
  sel_var = NA,
  by_cruise = FALSE,
  suffix = "",
  fig_res = 600
)
}
\arguments{
\item{sel_paths}{File paths to loocv output files}

\item{y_lab}{Y axis label as a character vector (i.e. "RSPE"). Default = NULL}

\item{make_plot}{Should a plot be generated}

\item{sel_var}{Character vector indicating what the variable name should be (used to name plot that is written to output)}

\item{by_cruise}{Logical. Should panels be divided by cruise? Default = FALSE plots panels by year.}

\item{suffix}{Text to append to a plot file name.}

\item{fig_res}{Resolution for the figure.}
}
\description{
Reads in output from leave-one-out cross-validation and makes a violin plot of prediction error and RMSPE. If make_plot is TRUE, writes a png file. Returns a ggplot object.
}
