% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_fermenter_custom.R
\name{scale_fill_fermenter_custom}
\alias{scale_fill_fermenter_custom}
\title{Custom fermenter fill scale for ggplot2}
\usage{
scale_fill_fermenter_custom(
  pal,
  na.value = "grey50",
  guide = "coloursteps",
  aesthetics = "fill",
  ...
)
}
\arguments{
\item{pal}{A character vector of colors to be used for the palette}

\item{na.value}{Color for NA values}

\item{guide}{Type of guide for ggplot2::binned_scale to use}

\item{aesthetics}{ggplot2 aesthetic type for the scale (e.g. "fill")}

\item{...}{Addditional arguments passed to ggplot2::binned_scale}
}
\description{
A fermenter scale that accepts custom color palettes.
}
