% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_timeseries_blue_strip}
\alias{theme_timeseries_blue_strip}
\title{Multi-panel map theme with blue strip}
\usage{
theme_timeseries_blue_strip()
}
\description{
Theme for multipanel maps with blue strip
}
