% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cold_pool_index}
\alias{cold_pool_index}
\title{Cold Pool Index (cold_pool_index)}
\format{
A data frame with 39 rows and 4 columns.
\describe{
     \item{YEAR}{Year}
     \item{COLD_POOL_AREA_KM2}{Cold Pool Index. Total area with bottom temperatures less than or equal to 2 celsius, in square kilometers}
     \item{LAST_UPDATE}{Date when cold pool index and temperature rasters were last updated.}
     \item{COLDPOOL_VERSION}{Version of the coldpool package that was used to calculate the cold pool index and produce temperature rasters}
}
}
\source{
\url{https://www.fisheries.noaa.gov/contact/groundfish-assessment-program}
}
\usage{
cold_pool_index
}
\description{
Data frame containing the cold pool index, defined as the area of the eastern Bering Sea continental shelf bottom trawl survey area with bottom temperatures less than or equal to two (2) degrees celsius during AFSC/RACE/GAP summer bottom trawl surveys.
}
\keyword{datasets}
