% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get gear temperature data from racebase and write to /data/}
\usage{
get_data(channel, include_preliminary_data = NULL)
}
\arguments{
\item{channel}{ODBC connection as an RODBC class}

\item{include_preliminary_data}{Character vector indicating whether to include preliminary data from the EBS ("ebs") and NBS ("nbs") Should the query return preliminary data from the current year from the race_data.edit_* tables? If this is used, the sql script needs to be updated for the current year.}
}
\description{
Get gear temperature data from racebase and write to /data/
}
