% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_raster_stack}
\alias{make_raster_stack}
\title{Make a raster stack from multiple rasters}
\usage{
make_raster_stack(file_path, file_name_contains = NULL, file_type = ".tif")
}
\arguments{
\item{file_path}{Path to the directory containing raster files}

\item{file_name_contains}{Character vector to filter on. (e.g., using "ste_" will ignore all filepaths that don't contain the characters sequence)}

\item{file_type}{File type. Default = ".tif" for GeoTIFF. Also works with native raster .grd and other stackable raster formats.}
}
\description{
Make a raster stack from multiple stackable raster layers (i.e. raster .grd, GeoTIFF, EHDR)
}
